requirements.txt
apex==0.9.10dev
boto3==1.18.37
botocore==1.21.37
easydict==1.9
flair==0.10
ftfy==6.0.3
gensim==3.8.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pytorch-crf==0.7.2
PyYAML==6.0
regex==2021.8.28
requests==2.25.1
scikit_learn==1.0.1
seqeval==1.2.2
spacy==3.2.0
tensorflow==1.14.1
torch==1.6.0+cu101
torchcrf==1.1.0
torchvision==0.7.0+cu101
tqdm==4.62.2
transformers==4.10.0
